requirements.txt
# Web Frameworks
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database
sqlalchemy==2.0.23

# API Clients
google-api-python-client==2.108.0
apify-client==1.7.1

# Web Scraping (backup)
beautifulsoup4==4.12.2

# Scheduling
apscheduler==3.10.4

# Logging
colorlog==6.8.0

# Dashboard (choose one)
flask==3.0.0  # OR
# fastapi==0.104.1
# uvicorn==0.24.0

# Data Visualization
plotly==5.18.0
